Users models with info, settings and filters v2
